Ensure Xcode developer dir is sert
I ran in to an issue in which the developer directory was set to something else and it caused an error in a playbook, stopping on the Accept License task. This would correct this issue and confirm that the correct directory is set. If the command fails, then the Xcode installation most likely failed and the playbook should stop anyway.